Move requests into client (#112)
* Move most of the requests into the client state itself.


* WIP

* Fix some errors and create new errors

* More teardown

* Fix detach requests

* Move set variable value to dap command

* Fix dap command error and add evaluate command

* FIx more compiler errors

* Fix more compiler errors

* Clipppyyyy

* FIx more

* One more

* Fix one more

* Use threadId from project instead u64

* Mostly fix stack frame list

* More compile errors

* More

* WIP transfer completions to dap command


* Finish console completions DapCommand impl

* Get Zed to build !!

* Fix test compile errors: The debugger tests will still fail

* Add threads reqeust to debug session


---------

+++ crates/dap/src/client.rs
@@ -27,11 +27,11 @@ const DAP_REQUEST_TIMEOUT: Duration = Duration::from_secs(12);
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[repr(transparent)]
-pub struct DebugAdapterClientId(pub usize);
+pub struct DebugAdapterClientId(pub u32);
 
 impl DebugAdapterClientId {
     pub fn from_proto(client_id: u64) -> Self {
-        Self(client_id as usize)
+        Self(client_id as u32)
     }
 
     pub fn to_proto(&self) -> u64 {
@@ -39,6 +39,7 @@ impl DebugAdapterClientId {
     }
 }
 
+/// Represents a connection to the debug adapter process, either via stdout/stdin or a socket.
 pub struct DebugAdapterClient {
     id: DebugAdapterClientId,
     sequence_count: AtomicU64,
